requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.9
google-play-scraper==1.2.5
openai>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
requests>=2.31.0
